+++ slsa-provenance/generator/action.yml
@@ -1,0 +1,17 @@
+name: "Provenance generator"
+description: "Generate provenance for a build in GitHub Actions"
+inputs:
+  digest:
+    description: "SHA256 digest of the binary built"
+    required: true
+  token:
+    description: 'GitHub API Token'
+    required: true
+outputs:
+  provenance:
+    description: "The JSON string of the provenance produced"
+runs:
+  using: "docker"
+  image: "Dockerfile"
+  env:
+    GITHUB_TOKEN: ${{ inputs.token }}
