+++ slsa-provenance/generator/go.mod
@@ -1,0 +1,11 @@
+module github.com/asraa/slsa-on-github
+
+go 1.16
+
+require (
+	github.com/google/go-github/v40 v40.0.0
+	github.com/in-toto/in-toto-golang v0.4.0-prerelease
+	github.com/shibumi/go-pathspec v1.3.0 // indirect
+	golang.org/x/oauth2 v0.0.0-20210402161424-2e8d93401602
+	golang.org/x/sys v0.0.0-20211124211545-fe61309f8881 // indirect
+)
